+++ src/Loader/FileLoader.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace AutoMapper\Loader;
 
-use AutoMapper\Generator\Generator;
+use AutoMapper\Generator\MapperGenerator;
 use AutoMapper\MapperGeneratorMetadataInterface;
 use PhpParser\PrettyPrinter\Standard;
 use PhpParser\PrettyPrinterAbstract;
@@ -20,7 +20,7 @@ final class FileLoader implements ClassLoaderInterface
     private ?array $registry = null;
 
     public function __construct(
-        private readonly Generator $generator,
+        private readonly MapperGenerator $generator,
         private readonly string $directory,
         private readonly bool $hotReload = true,
     ) {